Add ability to use Slack attachments
First quick approach to enable Slack attachments as payload:

If the message is a string, it is posted as 'text'. If the message is an array, it is posted as 'attachments'.

It should be an array of arrays: each main element is an attachment, and each attachment has attributes within it.

See https://api.slack.com/docs/attachments